+++ pkg/apis/core/v1/helper/helpers_test.go
@@ -22,7 +22,6 @@ import (
 	"testing"
 
 	v1 "k8s.io/api/core/v1"
-	apiequality "k8s.io/apimachinery/pkg/api/equality"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
@@ -243,70 +242,6 @@ func TestRemoveDuplicateAccessModes(t *testing.T) {
 	modes = removeDuplicateAccessModes(modes)
 	if len(modes) != 2 {
 		t.Errorf("Expected 2 distinct modes in set but found %v", len(modes))
-	}
-}
-
-func TestNodeSelectorRequirementsAsSelector(t *testing.T) {
-	matchExpressions := []v1.NodeSelectorRequirement{{
-		Key:      "foo",
-		Operator: v1.NodeSelectorOpIn,
-		Values:   []string{"bar", "baz"},
-	}}
-	mustParse := func(s string) labels.Selector {
-		out, e := labels.Parse(s)
-		if e != nil {
-			panic(e)
-		}
-		return out
-	}
-	tc := []struct {
-		in        []v1.NodeSelectorRequirement
-		out       labels.Selector
-		expectErr bool
-	}{
-		{in: nil, out: labels.Nothing()},
-		{in: []v1.NodeSelectorRequirement{}, out: labels.Nothing()},
-		{
-			in:  matchExpressions,
-			out: mustParse("foo in (baz,bar)"),
-		},
-		{
-			in: []v1.NodeSelectorRequirement{{
-				Key:      "foo",
-				Operator: v1.NodeSelectorOpExists,
-				Values:   []string{"bar", "baz"},
-			}},
-			expectErr: true,
-		},
-		{
-			in: []v1.NodeSelectorRequirement{{
-				Key:      "foo",
-				Operator: v1.NodeSelectorOpGt,
-				Values:   []string{"1"},
-			}},
-			out: mustParse("foo>1"),
-		},
-		{
-			in: []v1.NodeSelectorRequirement{{
-				Key:      "bar",
-				Operator: v1.NodeSelectorOpLt,
-				Values:   []string{"7"},
-			}},
-			out: mustParse("bar<7"),
-		},
-	}
-
-	for i, tc := range tc {
-		out, err := nodeSelectorRequirementsAsSelector(tc.in)
-		if err == nil && tc.expectErr {
-			t.Errorf("[%v]expected error but got none.", i)
-		}
-		if err != nil && !tc.expectErr {
-			t.Errorf("[%v]did not expect error but got: %v", i, err)
-		}
-		if !reflect.DeepEqual(out, tc.out) {
-			t.Errorf("[%v]expected:\n\t%+v\nbut got:\n\t%+v", i, tc.out, out)
-		}
 	}
 }
 
@@ -588,492 +523,6 @@ func TestGetAvoidPodsFromNode(t *testing.T) {
 		if !reflect.DeepEqual(tc.expectValue, v) {
 			t.Errorf("[%v]expect value %v but got %v with %v", i, tc.expectValue, v, v.PreferAvoidPods[0].PodSignature.PodController.Controller)
 		}
-	}
-}
-
-func TestMatchNodeSelectorTerms(t *testing.T) {
-	type args struct {
-		nodeSelector *v1.NodeSelector
-		node         *v1.Node
-	}
-
-	tests := []struct {
-		name string
-		args args
-		want bool
-	}{
-		{
-			name: "nil terms",
-			args: args{
-				nodeSelector: nil,
-				node:         nil,
-			},
-			want: false,
-		},
-		{
-			name: "node label matches matchExpressions terms",
-			args: args{
-				nodeSelector: &v1.NodeSelector{NodeSelectorTerms: []v1.NodeSelectorTerm{
-					{
-						MatchExpressions: []v1.NodeSelectorRequirement{{
-							Key:      "label_1",
-							Operator: v1.NodeSelectorOpIn,
-							Values:   []string{"label_1_val"},
-						}},
-					},
-				}},
-				node: &v1.Node{ObjectMeta: metav1.ObjectMeta{Labels: map[string]string{"label_1": "label_1_val"}}},
-			},
-			want: true,
-		},
-		{
-			name: "node field matches matchFields terms",
-			args: args{
-				nodeSelector: &v1.NodeSelector{NodeSelectorTerms: []v1.NodeSelectorTerm{
-					{
-						MatchFields: []v1.NodeSelectorRequirement{{
-							Key:      "metadata.name",
-							Operator: v1.NodeSelectorOpIn,
-							Values:   []string{"host_1"},
-						}},
-					},
-				}},
-				node: &v1.Node{ObjectMeta: metav1.ObjectMeta{Name: "host_1"}},
-			},
-			want: true,
-		},
-		{
-			name: "invalid node field requirement",
-			args: args{
-				nodeSelector: &v1.NodeSelector{NodeSelectorTerms: []v1.NodeSelectorTerm{
-					{
-						MatchFields: []v1.NodeSelectorRequirement{{
-							Key:      "metadata.name",
-							Operator: v1.NodeSelectorOpIn,
-							Values:   []string{"host_1, host_2"},
-						}},
-					},
-				}},
-				node: &v1.Node{ObjectMeta: metav1.ObjectMeta{Name: "host_1"}},
-			},
-			want: false,
-		},
-		{
-			name: "fieldSelectorTerm with node labels",
-			args: args{
-				nodeSelector: &v1.NodeSelector{NodeSelectorTerms: []v1.NodeSelectorTerm{
-					{
-						MatchFields: []v1.NodeSelectorRequirement{{
-							Key:      "metadata.name",
-							Operator: v1.NodeSelectorOpIn,
-							Values:   []string{"host_1"},
-						}},
-					},
-				}},
-				node: &v1.Node{ObjectMeta: metav1.ObjectMeta{Name: "not_host_1", Labels: map[string]string{
-					"metadata.name": "host_1",
-				}}},
-			},
-			want: false,
-		},
-		{
-			name: "labelSelectorTerm with node fields",
-			args: args{
-				nodeSelector: &v1.NodeSelector{NodeSelectorTerms: []v1.NodeSelectorTerm{
-					{
-						MatchExpressions: []v1.NodeSelectorRequirement{{
-							Key:      "metadata.name",
-							Operator: v1.NodeSelectorOpIn,
-							Values:   []string{"host_1"},
-						}},
-					},
-				}},
-				node: &v1.Node{ObjectMeta: metav1.ObjectMeta{Name: "host_1"}},
-			},
-			want: false,
-		},
-		{
-			name: "labelSelectorTerm and fieldSelectorTerm was set, but only node fields",
-			args: args{
-				nodeSelector: &v1.NodeSelector{NodeSelectorTerms: []v1.NodeSelectorTerm{
-					{
-						MatchExpressions: []v1.NodeSelectorRequirement{{
-							Key:      "label_1",
-							Operator: v1.NodeSelectorOpIn,
-							Values:   []string{"label_1_val"},
-						}},
-						MatchFields: []v1.NodeSelectorRequirement{{
-							Key:      "metadata.name",
-							Operator: v1.NodeSelectorOpIn,
-							Values:   []string{"host_1"},
-						}},
-					},
-				}},
-				node: &v1.Node{ObjectMeta: metav1.ObjectMeta{Name: "host_1"}},
-			},
-			want: false,
-		},
-		{
-			name: "labelSelectorTerm and fieldSelectorTerm was set, both node fields and labels (both matched)",
-			args: args{
-				nodeSelector: &v1.NodeSelector{NodeSelectorTerms: []v1.NodeSelectorTerm{
-					{
-						MatchExpressions: []v1.NodeSelectorRequirement{{
-							Key:      "label_1",
-							Operator: v1.NodeSelectorOpIn,
-							Values:   []string{"label_1_val"},
-						}},
-						MatchFields: []v1.NodeSelectorRequirement{{
-							Key:      "metadata.name",
-							Operator: v1.NodeSelectorOpIn,
-							Values:   []string{"host_1"},
-						}},
-					}},
-				},
-				node: &v1.Node{ObjectMeta: metav1.ObjectMeta{Name: "host_1",
-					Labels: map[string]string{
-						"label_1": "label_1_val",
-					}}},
-			},
-			want: true,
-		},
-		{
-			name: "labelSelectorTerm and fieldSelectorTerm was set, both node fields and labels (one mismatched)",
-			args: args{
-				nodeSelector: &v1.NodeSelector{NodeSelectorTerms: []v1.NodeSelectorTerm{
-					{
-						MatchExpressions: []v1.NodeSelectorRequirement{{
-							Key:      "label_1",
-							Operator: v1.NodeSelectorOpIn,
-							Values:   []string{"label_1_val"},
-						}},
-						MatchFields: []v1.NodeSelectorRequirement{{
-							Key:      "metadata.name",
-							Operator: v1.NodeSelectorOpIn,
-							Values:   []string{"host_1"},
-						}},
-					},
-				}},
-				node: &v1.Node{ObjectMeta: metav1.ObjectMeta{Name: "host_1",
-					Labels: map[string]string{
-						"label_1": "label_1_val-failed",
-					}}},
-			},
-			want: false,
-		},
-		{
-			name: "multi-selector was set, both node fields and labels (one mismatched)",
-			args: args{
-				nodeSelector: &v1.NodeSelector{NodeSelectorTerms: []v1.NodeSelectorTerm{
-					{
-						MatchExpressions: []v1.NodeSelectorRequirement{{
-							Key:      "label_1",
-							Operator: v1.NodeSelectorOpIn,
-							Values:   []string{"label_1_val"},
-						}},
-					},
-					{
-						MatchFields: []v1.NodeSelectorRequirement{{
-							Key:      "metadata.name",
-							Operator: v1.NodeSelectorOpIn,
-							Values:   []string{"host_1"},
-						}},
-					},
-				}},
-				node: &v1.Node{ObjectMeta: metav1.ObjectMeta{Name: "host_1",
-					Labels: map[string]string{
-						"label_1": "label_1_val-failed",
-					}}},
-			},
-			want: true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got, _ := MatchNodeSelectorTerms(tt.args.node, tt.args.nodeSelector); got != tt.want {
-				t.Errorf("MatchNodeSelectorTermsORed() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-// TestMatchNodeSelectorTermsStateless ensures MatchNodeSelectorTerms()
-// is invoked in a "stateless" manner, i.e. nodeSelector should NOT
-// be deeply modified after invoking
-func TestMatchNodeSelectorTermsStateless(t *testing.T) {
-	type args struct {
-		nodeSelector *v1.NodeSelector
-		node         *v1.Node
-	}
-
-	tests := []struct {
-		name string
-		args args
-		want *v1.NodeSelector
-	}{
-		{
-			name: "nil terms",
-			args: args{
-				nodeSelector: nil,
-				node:         nil,
-			},
-			want: nil,
-		},
-		{
-			name: "nodeLabels: preordered matchExpressions and nil matchFields",
-			args: args{
-				nodeSelector: &v1.NodeSelector{NodeSelectorTerms: []v1.NodeSelectorTerm{
-					{
-						MatchExpressions: []v1.NodeSelectorRequirement{{
-							Key:      "label_1",
-							Operator: v1.NodeSelectorOpIn,
-							Values:   []string{"label_1_val", "label_2_val"},
-						}},
-					},
-				}},
-				node: &v1.Node{ObjectMeta: metav1.ObjectMeta{Labels: map[string]string{"label_1": "label_1_val"}}},
-			},
-			want: &v1.NodeSelector{NodeSelectorTerms: []v1.NodeSelectorTerm{
-				{
-					MatchExpressions: []v1.NodeSelectorRequirement{{
-						Key:      "label_1",
-						Operator: v1.NodeSelectorOpIn,
-						Values:   []string{"label_1_val", "label_2_val"},
-					}},
-				},
-			}},
-		},
-		{
-			name: "nodeLabels: unordered matchExpressions and nil matchFields",
-			args: args{
-				nodeSelector: &v1.NodeSelector{NodeSelectorTerms: []v1.NodeSelectorTerm{
-					{
-						MatchExpressions: []v1.NodeSelectorRequirement{{
-							Key:      "label_1",
-							Operator: v1.NodeSelectorOpIn,
-							Values:   []string{"label_2_val", "label_1_val"},
-						}},
-					},
-				}},
-				node: &v1.Node{ObjectMeta: metav1.ObjectMeta{Labels: map[string]string{"label_1": "label_1_val"}}},
-			},
-			want: &v1.NodeSelector{NodeSelectorTerms: []v1.NodeSelectorTerm{
-				{
-					MatchExpressions: []v1.NodeSelectorRequirement{{
-						Key:      "label_1",
-						Operator: v1.NodeSelectorOpIn,
-						Values:   []string{"label_2_val", "label_1_val"},
-					}},
-				},
-			}},
-		},
-		{
-			name: "nodeFields: nil matchExpressions and preordered matchFields",
-			args: args{
-				nodeSelector: &v1.NodeSelector{NodeSelectorTerms: []v1.NodeSelectorTerm{
-					{
-						MatchFields: []v1.NodeSelectorRequirement{{
-							Key:      "metadata.name",
-							Operator: v1.NodeSelectorOpIn,
-							Values:   []string{"host_1", "host_2"},
-						}},
-					},
-				}},
-				node: &v1.Node{ObjectMeta: metav1.ObjectMeta{Name: "host_1"}},
-			},
-			want: &v1.NodeSelector{NodeSelectorTerms: []v1.NodeSelectorTerm{
-				{
-					MatchFields: []v1.NodeSelectorRequirement{{
-						Key:      "metadata.name",
-						Operator: v1.NodeSelectorOpIn,
-						Values:   []string{"host_1", "host_2"},
-					}},
-				},
-			}},
-		},
-		{
-			name: "nodeFields: nil matchExpressions and unordered matchFields",
-			args: args{
-				nodeSelector: &v1.NodeSelector{NodeSelectorTerms: []v1.NodeSelectorTerm{
-					{
-						MatchFields: []v1.NodeSelectorRequirement{{
-							Key:      "metadata.name",
-							Operator: v1.NodeSelectorOpIn,
-							Values:   []string{"host_2", "host_1"},
-						}},
-					},
-				}},
-				node: &v1.Node{ObjectMeta: metav1.ObjectMeta{Name: "host_1"}},
-			},
-			want: &v1.NodeSelector{NodeSelectorTerms: []v1.NodeSelectorTerm{
-				{
-					MatchFields: []v1.NodeSelectorRequirement{{
-						Key:      "metadata.name",
-						Operator: v1.NodeSelectorOpIn,
-						Values:   []string{"host_2", "host_1"},
-					}},
-				},
-			}},
-		},
-		{
-			name: "nodeLabels and nodeFields: ordered matchExpressions and ordered matchFields",
-			args: args{
-				nodeSelector: &v1.NodeSelector{NodeSelectorTerms: []v1.NodeSelectorTerm{
-					{
-						MatchExpressions: []v1.NodeSelectorRequirement{{
-							Key:      "label_1",
-							Operator: v1.NodeSelectorOpIn,
-							Values:   []string{"label_1_val", "label_2_val"},
-						}},
-						MatchFields: []v1.NodeSelectorRequirement{{
-							Key:      "metadata.name",
-							Operator: v1.NodeSelectorOpIn,
-							Values:   []string{"host_1", "host_2"},
-						}},
-					},
-				}},
-				node: &v1.Node{ObjectMeta: metav1.ObjectMeta{Name: "host_1",
-					Labels: map[string]string{
-						"label_1": "label_1_val",
-					}}},
-			},
-			want: &v1.NodeSelector{NodeSelectorTerms: []v1.NodeSelectorTerm{
-				{
-					MatchExpressions: []v1.NodeSelectorRequirement{{
-						Key:      "label_1",
-						Operator: v1.NodeSelectorOpIn,
-						Values:   []string{"label_1_val", "label_2_val"},
-					}},
-					MatchFields: []v1.NodeSelectorRequirement{{
-						Key:      "metadata.name",
-						Operator: v1.NodeSelectorOpIn,
-						Values:   []string{"host_1", "host_2"},
-					}},
-				},
-			}},
-		},
-		{
-			name: "nodeLabels and nodeFields: ordered matchExpressions and unordered matchFields",
-			args: args{
-				nodeSelector: &v1.NodeSelector{NodeSelectorTerms: []v1.NodeSelectorTerm{
-					{
-						MatchExpressions: []v1.NodeSelectorRequirement{{
-							Key:      "label_1",
-							Operator: v1.NodeSelectorOpIn,
-							Values:   []string{"label_1_val", "label_2_val"},
-						}},
-						MatchFields: []v1.NodeSelectorRequirement{{
-							Key:      "metadata.name",
-							Operator: v1.NodeSelectorOpIn,
-							Values:   []string{"host_2", "host_1"},
-						}},
-					},
-				}},
-				node: &v1.Node{ObjectMeta: metav1.ObjectMeta{Name: "host_1",
-					Labels: map[string]string{
-						"label_1": "label_1_val",
-					}}},
-			},
-			want: &v1.NodeSelector{NodeSelectorTerms: []v1.NodeSelectorTerm{
-				{
-					MatchExpressions: []v1.NodeSelectorRequirement{{
-						Key:      "label_1",
-						Operator: v1.NodeSelectorOpIn,
-						Values:   []string{"label_1_val", "label_2_val"},
-					}},
-					MatchFields: []v1.NodeSelectorRequirement{{
-						Key:      "metadata.name",
-						Operator: v1.NodeSelectorOpIn,
-						Values:   []string{"host_2", "host_1"},
-					}},
-				},
-			}},
-		},
-		{
-			name: "nodeLabels and nodeFields: unordered matchExpressions and ordered matchFields",
-			args: args{
-				nodeSelector: &v1.NodeSelector{NodeSelectorTerms: []v1.NodeSelectorTerm{
-					{
-						MatchExpressions: []v1.NodeSelectorRequirement{{
-							Key:      "label_1",
-							Operator: v1.NodeSelectorOpIn,
-							Values:   []string{"label_2_val", "label_1_val"},
-						}},
-						MatchFields: []v1.NodeSelectorRequirement{{
-							Key:      "metadata.name",
-							Operator: v1.NodeSelectorOpIn,
-							Values:   []string{"host_1", "host_2"},
-						}},
-					},
-				}},
-				node: &v1.Node{ObjectMeta: metav1.ObjectMeta{Name: "host_1",
-					Labels: map[string]string{
-						"label_1": "label_1_val",
-					}}},
-			},
-			want: &v1.NodeSelector{NodeSelectorTerms: []v1.NodeSelectorTerm{
-				{
-					MatchExpressions: []v1.NodeSelectorRequirement{{
-						Key:      "label_1",
-						Operator: v1.NodeSelectorOpIn,
-						Values:   []string{"label_2_val", "label_1_val"},
-					}},
-					MatchFields: []v1.NodeSelectorRequirement{{
-						Key:      "metadata.name",
-						Operator: v1.NodeSelectorOpIn,
-						Values:   []string{"host_1", "host_2"},
-					}},
-				},
-			}},
-		},
-		{
-			name: "nodeLabels and nodeFields: unordered matchExpressions and unordered matchFields",
-			args: args{
-				nodeSelector: &v1.NodeSelector{NodeSelectorTerms: []v1.NodeSelectorTerm{
-					{
-						MatchExpressions: []v1.NodeSelectorRequirement{{
-							Key:      "label_1",
-							Operator: v1.NodeSelectorOpIn,
-							Values:   []string{"label_2_val", "label_1_val"},
-						}},
-						MatchFields: []v1.NodeSelectorRequirement{{
-							Key:      "metadata.name",
-							Operator: v1.NodeSelectorOpIn,
-							Values:   []string{"host_2", "host_1"},
-						}},
-					},
-				}},
-				node: &v1.Node{ObjectMeta: metav1.ObjectMeta{Name: "host_1",
-					Labels: map[string]string{
-						"label_1": "label_1_val",
-					}}},
-			},
-			want: &v1.NodeSelector{NodeSelectorTerms: []v1.NodeSelectorTerm{
-				{
-					MatchExpressions: []v1.NodeSelectorRequirement{{
-						Key:      "label_1",
-						Operator: v1.NodeSelectorOpIn,
-						Values:   []string{"label_2_val", "label_1_val"},
-					}},
-					MatchFields: []v1.NodeSelectorRequirement{{
-						Key:      "metadata.name",
-						Operator: v1.NodeSelectorOpIn,
-						Values:   []string{"host_2", "host_1"},
-					}},
-				},
-			}},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			MatchNodeSelectorTerms(tt.args.node, tt.args.nodeSelector)
-			if !apiequality.Semantic.DeepEqual(tt.args.nodeSelector, tt.want) {
-				// fail when tt.args.nodeSelector is deeply modified
-				t.Errorf("MatchNodeSelectorTerms() got = %v, want %v", tt.args.nodeSelector, tt.want)
-			}
-		})
 	}
 }
