+++ pkg/scheduler/framework/plugins/helper/node_affinity.go
@@ -19,7 +19,7 @@ package helper
 import (
 	"k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/labels"
-	v1helper "k8s.io/kubernetes/pkg/apis/core/v1/helper"
+	"k8s.io/component-helpers/scheduling/corev1"
 )
 
 // PodMatchesNodeSelectorAndAffinityTerms checks whether the pod is schedulable onto nodes according to
@@ -71,6 +71,6 @@ func PodMatchesNodeSelectorAndAffinityTerms(pod *v1.Pod, node *v1.Node) bool {
 // terms are ORed, and an empty list of terms will match nothing.
 func nodeMatchesNodeSelectorTerms(node *v1.Node, nodeSelector *v1.NodeSelector) bool {
 	// TODO(@alculquicondor, #95738): parse this error earlier in the plugin so we only need to do it once per pod
-	matches, _ := v1helper.MatchNodeSelectorTerms(node, nodeSelector)
+	matches, _ := corev1.MatchNodeSelectorTerms(node, nodeSelector)
 	return matches
 }
